Fishing base mechanics finished

+++ Assets/Minigames/Fish/Scripts/Settings/FishSpawnSettings.cs
@@ -11,9 +11,10 @@ namespace Minigames.Fish
     [Serializable]
     public class FishSpawnSettings : ScriptableObject
     {
-        public float DepthInterval;
+        public int DepthInterval;
         public int FishPerWave;
         public float SpawnAreaWidth;
+
     }
 
 }
